requirements.txt
langchain==0.1.4
streamlit==1.30.0
langchain-openai==0.0.5
langchain-core==0.1.18
langchain-community==0.0.16  
beautifulsoup4==4.12.3
python-dotenv==1.0.1
chromadb==0.4.22
